% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_param_2022.R
\name{build.param}
\alias{build.param}
\title{Script replacement for fmat function in mmage package.}
\usage{
build.param(
  tcla,
  parms,
  phase,
  offtake,
  female.offtake = NULL,
  prolificacyRate = NULL,
  fbirthRate = 0.5,
  lambda = 1,
  correctionfec = TRUE,
  truncated = TRUE,
  method = "steady",
  hazards = FALSE
)
}
\arguments{
\item{tcla}{data.frame created with fclass function (see mmage package)}

\item{parms}{a list containing:
pfbirth (vector of probabilities of each age class giving birth to a female),
part.age (age of first parturition; index corresponding to class in tcla$class ),
prolificacy (average number of offspring per birth per age, corresponding to age classes)
f.mortality (a vector of length=number of age classes specifying mortality rates for each class of females)
m.mortality (a vector of length=number of age classes specifying mortality rates for each class of males)
f.offtake (a vector of length =number of age classes specifying culling rates for each class of females)}

\item{phase}{a string either "year" or "month". To model annual growth, nbphase must be set to 1 when creating tcla and nbphase set to "year"
To model monthly changes in growth set nbphase to "month" and 12 when creating tcla table.}

\item{offtake}{a vector of length=length of age classes giving the harvest profile (probability of individual being killed under a particular strategy)}

\item{lambda}{set to 1 default}

\item{correctionfec}{default is TRUE}

\item{method}{set to "steady"}

\item{hazards}{set to TRUE if output should include a table of hazards rather than probabilities (i.e., variable hh in code)}

\item{parturition}{parturition rate - a number representing frequency of births per year}
}
\value{
a data.frame to use for projecting herd demography
}
\description{
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
}
\examples{

# Build tcla table
lclass = c(0.17, 0.5, 1, 2, 3, 4, 5, 6, 7)
tcla = build.tcla(female.ages = lclass, male.ages = lclass, nbphase = 12)

# build parms list
parms = list(
  parturition = 1.2,
  part.age = 2,  # age of first parturition
# number of livebirths expected per animal per parturition per year by age class
  prolificacy = c( 0, 0, 0, 0.82, 1.10, 1.41, 1.45, 1.03, 1.03),
  f.mortality = c( 0.05, 0.45, 0.18, 0.18, 0.18, 0.18, 0.18,  0.18,  0.18),
# m.mortality = buck.mortality
# f.offtake = doe.mortality/2
)

# run function
build.param = function(tcla, parms, phase="month", offtake=c(0.45,  0.45, 0.15, 0.15, 0.05, 0.05, 0.05,  0.05,  1.00))

}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
