% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTCLA.R
\name{build.tcla}
\alias{build.tcla}
\title{Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par}
\usage{
build.tcla(female.ages, male.ages, nbphase)
}
\arguments{
\item{female.ages}{a list containing age classes of females}

\item{male.ages}{a list containing age classes of males}

\item{nbphase}{an integer by which age classes will be multiplied. Set nbphase to 1 to model population growth by year, set to 12 to model monthly growth}
}
\value{
data.frame created with fclass function (see mmage package)
}
\description{
Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
