% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildparam_advanced.R
\name{build_param}
\alias{build_param}
\title{Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par}
\usage{
build_param(
  tcla,
  parms,
  nbphase,
  offtake,
  female.offtake = NULL,
  phi = NULL,
  Inf.Mortality = c(NULL, "auto"),
  prolificacyRate = c(NULL, "auto"),
  fbirthRate = 0.5,
  lambda = 1,
  correctionfec = TRUE,
  method = "steady",
  terminal = c("off", "dea"),
  hazards = FALSE
)
}
\arguments{
\item{tcla}{data.frame created with fclass function (see mmage package)}

\item{parms}{a list containing:
\itemize{
\item pfbirth (vector of probabilities of each age class giving birth to a female),
\item part.age (age of first parturition; index corresponding to class in tcla$class ),
\item prolificacy (average number of offspring per birth per age, corresponding to age classes)
\item f.mortality (a vector of length=number of age classes specifying mortality rates for each class of females)
\item m.mortality (a vector of length=number of age classes specifying mortality rates for each class of males)
\item f.offtake (a vector of length=number of age classes specifying culling rates for each class of females)
}}

\item{nbphase}{an integer, set as 1 for yearly projections, 12 for monthly.}

\item{offtake}{a vector of length=length of age classes giving the harvest profile (probability of individual being killed under a particular strategy)}

\item{female.offtake}{Integer. Specifies the factor by which offtake rates should be reduced for female animals. If left as the default, NULL, the same offtake rates are applied to both males and females. Can be a vector of probabilities with length equal to length(lclass).}

\item{phi}{for use in optimizing routines. Female offtake probabilities will be multiplied by phi if a value is specified. Default is 1.}

\item{Inf.Mortality}{if set to "auto", variation in mortality for the first three age classes is introduced using normal distribution of mean and sd obtained from input \emph{parms$f.mortality}}

\item{prolificacyRate}{if set to "auto", variation in prolificacy for the reproductive age classes is introduced using normal distribution of mean and sd obtained from input \emph{parms$MeanProlificacy} and \emph{parms$sdProlificacy}}

\item{fbirthRate}{default is 0.5 (probability of a female birth)}

\item{lambda}{set to 1 default}

\item{correctionfec}{default is TRUE}

\item{method}{set to "steady"}

\item{terminal}{Set to "off" if probability of offtake competes with probability of intrinsic mortality, otherwise set to "dea". This is only used for truncated models where terminal age class is set to 1.}

\item{hazards}{set to TRUE if output should include a table of hazards rather than probabilities (i.e., variable hh in code)}
}
\value{
a data.frame to use for projecting herd demography
}
\description{
Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
}
\examples{
  lclass <- c(0.17, 0.5, 1, 1, 1, Inf)
  nbphase = 12
  tcla <- build_tcla( female.ages = lclass, male.ages = lclass, nbphase = nbphase)
  parms = list(
    ages = c( 0.17, 0.5, 1, 2, 3, 4),
    parturition = 1.2,
    part.age = 2,
    prolificacy = c( 0, 0.82, 1.10, 1.41, 1.45, 1.03),
    f.mortality = c(0, 0.10, 0.10, 0.20, 0.50, 0.75),
    m.mortality = c(0, 0.10, 0.10, 0.20, 0.50, 0.75)
    )
  build_param(tcla, parms=parms, nbphase = nbphase, offtake = c(0, 0, 0.10, 0.25, 0.50, 0.75, 0.90 ), prolificacyRate = "auto", Inf.Mortality = "auto")
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
