% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_param.R
\name{build_param_depricated}
\alias{build_param_depricated}
\title{This function builds the param table containing age and sex classes needed to project a population using mmage package
Depricated, to be replaced by advanced build}
\usage{
build_param_depricated(
  tcla,
  parms,
  phase,
  offtake,
  male.offtake = NULL,
  prolificacyRate = NULL,
  fbirthRate = 0.5,
  lambda = 1,
  correctionfec = TRUE,
  truncated = TRUE,
  method = "steady",
  hazards = FALSE
)
}
\arguments{
\item{tcla}{A data.frame created with mmage::fclass().}

\item{parms}{A list containing:
ages (a vector of ages equal the length of lclass in tcla table)
pfbirth (vector of probabilities of each age class giving birth to a female),
part.age (age of first parturition; index corresponding to class in tcla$class ),
prolificacy (average number of offspring per birth per age, corresponding to age classes)
f.mortality (a vector of length=number of age classes specifying mortality rates for each class of females)
m.mortality (a vector of length=number of age classes specifying mortality rates for each class of males) and
male.offtake (a vector of length =number of age classes specifying culling rates for each class of females).}

\item{phase}{A string either "year" or "month". To model annual growth, nbphase must be set to 1 when creating tcla and nbphase set to "year"
To model monthly changes in growth set nbphase to "month" and 12 when creating tcla table.}

\item{offtake}{A vector of length=length of age classes giving the harvest profile (probability of individual being killed under a particular strategy).}

\item{male.offtake}{A vector of length=length of age classes giving the culling rates for males. If NULL (default) the culling rates in 'offtake' are used for both males and females.}

\item{lambda}{Set to 1 by default as the growth rate of the herd.}

\item{correctionfec}{Boolean, if False, a parameter set for birth pulse is generated. If true (default), birth pulse parameter set.}

\item{method}{Set to "steady", which is used to simulate birth flow. Alternative is "geom" used to simulate birth pulse.}

\item{hazards}{Set to TRUE if output should include a table of hazards rather than probabilities (i.e., variable hh in code)}

\item{parturition}{A number representing frequency of births per year, otherwise known as the parturition rate.}
}
\value{
A data.frame to use for projecting herd demography.
}
\description{
This function builds the param table containing age and sex classes needed to project a population using mmage package
Depricated, to be replaced by advanced build
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
##Old example##
lclass <- c(0.17, 0.5, 1, 1, 1, 1)
tcla <- build_tcla( female.ages = lclass, male.ages = lclass, nbphase = 12)
parms = list(
ages = c( 0.17, 0.5, 1, 2, 3, 4),
parturition = 1.2,
part.age = 2,
prolificacy = c( 0, 0.82, 1.10, 1.41, 1.45, 1.03),
f.mortality = c(0, 0.10, 0.10, 0.20, 0.50, 0.75),
m.mortality = c(0, 0.10, 0.10, 0.20, 0.50, 0.75)
)
build_param_depricated(tcla, parms=parms, phase = "month", offtake = c(0, 0, 0.10, 0.25, 0.50, 0.75, 0.90 ), male.offtake = NULL, prolificacyRate = NULL, fbirthRate = 0.5, lambda = 1, correctionfec = TRUE, truncated = TRUE, method = "steady", hazards = FALSE)
}
