% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/culling_multiplot.R
\name{culling_multiplot}
\alias{culling_multiplot}
\title{Quick multi-plot of survivorship curves based on offtake rates}
\usage{
culling_multiplot(
  survivorship_profiles = offtake_models,
  Payne = TRUE,
  agecats = NULL
)
}
\arguments{
\item{survivorship_profiles}{a named list of vectors with numbers corresponding the offtake probabilities by age. Default is offtake_models list included with HerdDynamics}

\item{Payne}{if TRUE (default) uses Payne's 1973 age class categories. Note: provided survivorship profiles must contain the same number (9) of survivorship probabilities as Payne's age classes. If FALSE, the user must provide their own age categories in the agecats argument.}

\item{agecats}{a vector of age categories to use for plotting. Length must match length of individual survivorship probabilities}
}
\value{
a line graph with offtake models colored by strategy.
}
\description{
Quick multi-plot of survivorship curves based on offtake rates
}
\examples{
  x = list(
    Meat = c(0, 0.10, 0.25, 0.50, 0.75, 0.90 ),
    Milk = c(0, 0.10, 0.15, 0.20, 0.65, 0.95 )
    )
  a = c(0.5, 1, 2, 3, 4, 5)
  plot_offtake(offtake = x, ages = a , title="Meat vs Milk")
}
\references{
Payne, S. (1973). Kill-off Patterns in Sheep and Goats: The Mandibles from Aşvan Kale. Anatolian Studies, 23, 281–303. https://doi.org/10.2307/3642547
}
