% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_prolificacy_rates.R
\name{generate_prolificacy_rates}
\alias{generate_prolificacy_rates}
\title{Generate Prolificacy rates
This function generates prolificacy rates based on mean and standard deviation of provided list of prolificacy rates}
\usage{
generate_prolificacy_rates(meanPro, sdPro, n, sample_size = 100)
}
\arguments{
\item{meanPro}{mean prolificacy}

\item{sdPro}{standard deviation of prolificacy}

\item{n}{number of prolificacy rates to generate}

\item{sample_size}{number of samples to generate}
}
\value{
a vector of generated prolificacy rates

#-- function to autogenerate prolificacy rates, used when prolificacyRate set to "auto". These lines to be removed in package version.
}
\description{
Generate Prolificacy rates
This function generates prolificacy rates based on mean and standard deviation of provided list of prolificacy rates
}
