% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLambda.R
\name{getLambda}
\alias{getLambda}
\title{Function for computing reproduction traits
This function takes param as input and returns lambda, reproductive value, and herd sex proportions}
\usage{
getLambda(param, tcla, lclass, lclassm = NULL, p0)
}
\arguments{
\item{param}{data.frame created with fvh2par function (see mmage package)}

\item{tcla}{dataframe of the original tcla table used to produce param (see mmage package)}

\item{lclass}{a vector of age classes (same as used to build tcla in earlier step)}

\item{lclassm}{a vector of age classes for males used to build tcla, if applicable.
Default is NULL, which assumes the same ages were used for males and females when building tcla.}

\item{p0}{integer the initial population size}
}
\value{
a list containing: lambda of herd, a dataframe with sex proportions of the herd, and a dataframe with initial herd traits, including reproductive value by sex/age class
}
\description{
Function for computing reproduction traits
This function takes param as input and returns lambda, reproductive value, and herd sex proportions
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
