% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerdDynamics-package.R
\docType{package}
\name{HerdDynamics-package}
\alias{HerdDynamics}
\alias{HerdDynamics-package}
\title{HerdDynamics: Enables Livestock Demographic Modelling}
\description{
HerdDynamics can be used to project age and sex structured population changes of livestock herds. Input parameters include age- and sex- specific offtake rates, fertility (i.e., parturitions and prolificacy), and mortality. Additional tools allow for adjustment factors to be generated to reach a steady state of herd growth.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/penguinnick/HerdDynamics}
  \item Report bugs at \url{https://github.com/penguinnick/HerdDynamics/issues}
}

}
\author{
\strong{Maintainer}: Nicholas Triozzi \email{triozzi@ucsb.edu} (\href{https://orcid.org/0000-0002-1856-7449}{ORCID})

}
\keyword{internal}
