% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmagefmat.R
\name{mmage.fmat}
\alias{mmage.fmat}
\title{Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
Build matrices F, D, O, I, G and A from the input parameters.}
\usage{
mmage.fmat(param, Lf, Lm)
}
\arguments{
\item{param}{dataframe built from birth, survival, and death rates}
}
\value{
list of arrays FEC, DEA, OFF, ID, G, A, and PAR
}
\description{
Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
Build matrices F, D, O, I, G and A from the input parameters.
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
