% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmagezcal.R
\name{mmage.zcal}
\alias{mmage.zcal}
\title{Rewritten zcal script from mmage updating aggregate function, eliminating formula param from aggregate function
This function creates the calendar matrix}
\usage{
mmage.zcal(nbphase, nbstep)
}
\arguments{
\item{nbphase}{number of phases (1 for yearly, 12 for monthly). Inherited}

\item{nbstep}{product of nbphase and nbcycle}
}
\description{
Rewritten zcal script from mmage updating aggregate function, eliminating formula param from aggregate function
This function creates the calendar matrix
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
