% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmage_flcass.R
\name{mmage_fclass}
\alias{mmage_fclass}
\title{A modification of mmage::flcass function used to build tcla table.}
\usage{
mmage_fclass(lclassf, lclassm = NULL)
}
\arguments{
\item{lclassf}{a vector giving the lengths of each age class for females}

\item{lclassm}{a vector giving the lengths of each age class for males}

\item{inf}{if TRUE}
}
\description{
A modification of mmage::flcass function used to build tcla table.
}
