% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offtake_to_df.R
\name{offtake_to_df}
\alias{offtake_to_df}
\title{Prepare list of offtake models for printing}
\usage{
offtake_to_df(offtake, ages)
}
\arguments{
\item{offtake}{a list of vectors with numbers corresponding the offtake probabilities by age}

\item{ages}{a vector of numbers corresponding to the age of an animal. Length should be equal to length of each offtake model}
}
\description{
Prepare list of offtake models for printing
}
