% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_list.R
\name{param_list}
\alias{param_list}
\title{bundles param and repro parameter sets into a list for use in \emph{project_herd} function}
\usage{
param_list(param, repro)
}
\arguments{
\item{param}{a dataframe or list of dataframes created with \emph{build_param} function}

\item{repro}{a list (or lists of) of vectors and dataframe created with \emph{get_lambda} function}
}
\value{
a list of parameters
}
\description{
bundles param and repro parameter sets into a list for use in \emph{project_herd} function
}
