% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Payne_ages}
\alias{Payne_ages}
\title{Payne's Age Classes}
\format{
\subsection{\code{Payne_ages}}{

A data frame with 9 rows and 3 columns:
\describe{
\item{ageClasses}{string/character class containing Age classes A-I}
\item{ages}{Age of animals in each age class in years}
\item{lclass}{the length of each age class in months}

}
}
}
\source{
Payne, S. (1973). Kill-off Patterns in Sheep and Goats: The Mandibles from Aşvan Kale. Anatolian Studies, 23, 281–303. https://doi.org/10.2307/3642547
}
\usage{
Payne_ages
}
\description{
A dataframe containing Payne's age classes for sheep and goats used to
initialize the tcla table.
}
\keyword{datasets}
