% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survivorship.R
\name{plot_survivorship}
\alias{plot_survivorship}
\title{Quickly plot survivorship curves}
\usage{
plot_survivorship(
  x,
  ages = NULL,
  title = "Survivorship Curves",
  key = "strategy"
)
}
\arguments{
\item{x}{a list of vectors with numbers corresponding the survival probabilities by age}

\item{ages}{a vector of numbers corresponding to the age of an animal. Length should be equal to length of each offtake model}

\item{title}{a string providing the title for the graph}

\item{key}{a string providing label to use in legend, e.g., "Strategy" or "Site"}
}
\value{
a line graph with offtake models colored by strategy.
}
\description{
Quickly plot survivorship curves
}
\examples{
plot_survivorship(
  x = list(
    Meat = c(0, 0.10, 0.25, 0.50, 0.75, 0.90),
    Milk = c(0, 0.10, 0.15, 0.20, 0.65, 0.95)
  ),
  ages = c(0.5, 1, 2, 3, 4, 5),
  title = "Meat vs Milk",
  key = "Strategy"
)
}
