% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_summary.R
\name{pop_summary}
\alias{pop_summary}
\title{Script summarizes population change through time.}
\usage{
pop_summary(x, interval)
}
\arguments{
\item{x}{list created with projectHerd function}

\item{interval}{a character string, either "month" to get results by time step or "year" to get results by cycle}
}
\value{
data.frame with two columns (time, population)
}
\description{
Script summarizes population change through time.
}
