% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res2df.R
\name{res2df}
\alias{res2df}
\title{transform projection results into a dataframe for plotting with ggplot}
\usage{
res2df(result.list)
}
\arguments{
\item{result.list}{a list of results produced using pop_summary function}
}
\value{
a dataframe containing time step, population size, and strategy
}
\description{
transform projection results into a dataframe for plotting with ggplot
}
