% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivorship.R
\name{survivorship}
\alias{survivorship}
\title{Survivorship function to calculate survivorship probabilities from absolute frequencies}
\usage{
survivorship(x)
}
\arguments{
\item{x}{is a vector containing the absolute frequencies of individuals in each age class}
}
\value{
Running the function will give a vector equal in length to x containing survival probabilities for each age class.
}
\description{
Survivorship function to calculate survivorship probabilities from absolute frequencies
}
\references{
Price, M., Wolfhagen, J. and Otárola-Castillo, E. (2016). Confidence Intervals in the Analysis of Mortality and Survivorship Curves in Zooarchaeology. Am. Antiq. 81: 157–173.
}
