% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_infant_mortality_rates.R
\name{generate_infant_mortality_rates}
\alias{generate_infant_mortality_rates}
\title{Generate Infant mortality rates
This function generates infant mortality rates rates based on mean and standard deviation of provided list of mortality rates.}
\usage{
generate_infant_mortality_rates(Mort, n = 3)
}
\arguments{
\item{Mort}{vector of mortality rates}

\item{n}{number of mortality rates corresponding to the number of infant age classes (ages below 1). Default is 3.}
}
\value{
a vector of generated prolificacy rates
}
\description{
Generate Infant mortality rates
This function generates infant mortality rates rates based on mean and standard deviation of provided list of mortality rates.
}
\section{Functions}{
\itemize{
\item \code{generate_infant_mortality_rates()}: Generate infant mortality rates

}}
