% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_param.R
\name{build_param}
\alias{build_param}
\title{Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par}
\usage{
build_param(
  tcla,
  parms,
  phase,
  offtake,
  male.offtake = FALSE,
  prolificacyRate = NULL,
  fbirthRate = 0.5,
  lambda = 1,
  correctionfec = TRUE,
  truncated = TRUE,
  method = "steady",
  hazards = FALSE
)
}
\arguments{
\item{tcla}{data.frame created with fclass function (see mmage package)}

\item{parms}{a list containing:
pfbirth (vector of probabilities of each age class giving birth to a female),
part.age (age of first parturition; index corresponding to class in tcla$class ),
prolificacy (average number of offspring per birth per age, corresponding to age classes)
f.mortality (a vector of length=number of age classes specifying mortality rates for each class of females)
m.mortality (a vector of length=number of age classes specifying mortality rates for each class of males)
f.offtake (a vector of length =number of age classes specifying culling rates for each class of females)}

\item{phase}{a string either "year" or "month". To model annual growth, nbphase must be set to 1 when creating tcla and nbphase set to "year"
To model monthly changes in growth set nbphase to "month" and 12 when creating tcla table.}

\item{offtake}{a vector of length=length of age classes giving the harvest profile (probability of individual being killed under a particular strategy)}

\item{lambda}{set to 1 default}

\item{correctionfec}{Boolean, if False, a parameter set for birth pulse is generated. If true (default), birth pulse parameter set}

\item{method}{set to "steady". Alternative is "geometric"}

\item{hazards}{set to TRUE if output should include a table of hazards rather than probabilities (i.e., variable hh in code)}

\item{parturition}{parturition rate - a number representing frequency of births per year}
}
\value{
a data.frame to use for projecting herd demography
}
\description{
Script replacement for fmat function in mmage package.
This function solves an error thrown when fmat(param) is run by specifying Lf and Lm as performed in fvh2par
}
\examples{
build_param( tcla, parms, phase = "month", offtake = c(0, 0.10, 0.25, 0.50, 0.75, 0.90 ))
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
