% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_tcla.R
\name{build_tcla}
\alias{build_tcla}
\title{Build age-class and sex structured table. The output table (using mmage naming convention tcla) is the first step for herd projection.}
\usage{
build_tcla(female.ages, male.ages = NULL, nbphase = 1)
}
\arguments{
\item{female.ages}{A numeric vector specifying the bin widths of age classes in years for females where c(0.5, 1, 1) corresponds to 6 months, 1 year, 2 years of age. If terminal age is not Inf, a truncated model will be produced.}

\item{male.ages}{a vector of numbers containing the length of age classes in years for males where c(0.5, 1, 1) corresponds to 6 months, 1 year, 2 years of age. If terminal age is not Inf, a truncated model will be produced.}

\item{nbphase}{An integer by which age classes will be multiplied. Default is 1. Set to 12 to model population growth by month.}
}
\value{
data.frame created with fclass function (see mmage package)
}
\description{
Build age-class and sex structured table. The output table (using mmage naming convention tcla) is the first step for herd projection.
}
\examples{
  build_tcla(female.ages = c(1, 3, 10, Inf), male.ages = c(1, 3, 5, Inf), nbphase = 1)
}
\references{
Lesnoff, M. (*), 2015. mmage: A R package for age-structured population matrix models. CIRAD, Montpellier, France. http://livtools.cirad.fr.
}
