% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_offtake.R
\name{plot_offtake}
\alias{plot_offtake}
\title{Quickly plot a list of offtake models}
\usage{
plot_offtake(offtake, ages, title = title)
}
\arguments{
\item{offtake}{a list of vectors with numbers corresponding the offtake probabilities by age}

\item{ages}{a vector of numbers corresponding to the age of an animal. Length should be equal to length of each offtake model}

\item{title}{a string providing the title for the graph}
}
\value{
a line graph with offtake models colored by strategy.
}
\description{
Quickly plot a list of offtake models
}
\examples{
plot.offtake(offtake = list( Meat = c(0, 0.10, 0.25, 0.50, 0.75, 0.90 ), Milk = c(0, 0.10, 0.15, 0.20, 0.65, 0.95 )), ages = c(0.5, 1, 2, 3, 4, 5), title="Meat vs Milk")
}
