% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_summary.R
\name{pop_summary}
\alias{pop_summary}
\title{Script summarizes population change through time.}
\usage{
pop_summary(x, ..., interval = NULL)
}
\arguments{
\item{x}{list created with projectHerd function}

\item{...}{field names passed on to group_by arg in dplyr. This will be any combination of tim, cycle, sex, class, etc.}
}
\value{
data.frame with two columns (time, population)
}
\description{
Script summarizes population change through time.
}
